feat: Firebase 초기화 및 알림 권한 요청 기능 추가
